moved all wrapper functions from cfuncs.py to _param_config.py
